+++ subscribe_test.go
@@ -47,7 +47,7 @@ func TestChainClient_Subscribe(t *testing.T) {
 				Amount: 10,
 			},
 			Timestamp: time.Now().UnixNano(),
-			Nonce:     rand.Int63(),
+			Nonce:     uint64(rand.Int63()),
 		}
 
 		err = tx.Sign(privKey)
